changed badgestyle
